Implemented EventContext for agent reactions
This commit implements the EventContext class for agent reactions.
It allows for the creation of a context object that can be used to
store the context of an agent reaction.

The EventContext class is a base class that can be used to create
a context object for an agent reaction.